services/swipes/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-cloud-firestore
python-dotenv
